+++ chat-components/src/components/footer/subcomponents/DownloadTranscriptButton.tsx
@@ -24,8 +24,8 @@ function DownloadTranscriptButton(props: ICommandButtonProps) {
             onClick={props.onClick}
             ariaLabel={props.ariaLabel ?? AriaLabels.DownloadChatTranscript}
             className={props.className} 
-            customEvent={customEvent}
-            hideButtonTitle ={props.hideButtonTitle}/>);
+            customEvent={customEvent}/>
+    );
 }
 
 export default DownloadTranscriptButton;