+++ chat-components/src/components/header/subcomponents/MinimizeButton.tsx
@@ -25,8 +25,7 @@ function MinimizeButton(props: ICommandButtonProps) {
             onClick={props.onClick}
             ariaLabel={props.ariaLabel ?? AriaLabels.Minimize}
             className={props.className}
-            customEvent={customEvent}
-            hideButtonTitle = {props.hideButtonTitle}/>
+            customEvent={customEvent}/>
     );
 }